New Features
- Added an explicit close button for dismissing side panels.
- Added an automatic name generation for importing individual
text files as wiki entries. The files' names shall be their
title.

This change does not affect importing wiki entries from JSON
data sets.

- Re-arranged the editing interface to have the following order:

1. Title
2. Tags
3. Content
4. Mime type

This is to be consistent with the viewing interface, which
displays wiki entry information in that order.

- Fixed an issue where clicking on the indexing button for the
first time forces the screen to scroll to bottom in Firefox.